pyproject.toml
[project]
name = "regulations-rag-eval"
version = "0.1.0"
description = "Evaluation framework for RAG implementations for working with regulations"
readme = "README.md"
authors = [
    { name = "jrwone0", email = "jrwilliams@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click",
    "pandas",
    "pydantic",
    "httpx",
    "openai",
    "faiss-cpu",
    "pypdf",
    "langchain",
    "langchain_core",
    "langchain_neo4j",
    "langchain_community",
    "langchain_openai",
    "huggingface_hub",
    "neo4j",
    "tenacity",
    "networkx",
    "chromadb",
    "numpy",
    "reportlab",
    "flask",
    "python-dotenv",  # Corrected from 'dotenv'
    "aiofiles",
    "weaviate-client",  # Added for ColBERT implementation
    "torch",  # Added for ColBERT implementation
    "transformers",  # Added for ColBERT implementation
    "pyyaml"  # Added for ColBERT implementation
]

[project.scripts]
regulations-rag-eval = "regulations_rag_eval:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = ["pytest>=8.0.0"]
